requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dnspython==2.6.1
Flask==3.0.3
flask-cors==4.0.1
fonttools==4.53.0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.7.1
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
Werkzeug==3.0.3
numpy==1.26.4
flask-jwt-extended==4.6.0
bcrypt==4.1.3
Flask-SQLAlchemy==3.1.1
psycopg[binary]==3.1.19
Flask-Limiter==3.7.0
pytest==8.2.2
marshmallow==3.21.1
pandas==2.2.3
alembic==1.14.0
pyotp==2.9.0
cryptography==43.0.1
gunicorn==23.0.0
